+++ scheme.go
@@ -1,0 +1,3 @@
+package scy
+
+import _ "github.com/viant/afsc/op"